requirements.txt
# Phase 1 — parser (GPX ingestion and classification)
pyyaml>=6.0

# Phase 2 — enrichment (weather, sheet join, petrol price, Bluelink, full pipeline)
requests>=2.31
hyundai-kia-connect-api>=4.10

# Phase 3 — outputs (heatmap, dashboard)
# pandas
# folium

# Phase 4 — clustering (route similarity labels)
# scikit-learn
